// frontend/src/generated/openapi/model/event.ts
/**
 * api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Employee } from './employee';
import { Team } from './team';


export interface Event { 
    details: string;
    employees: Array<Employee>;
    end: string;
    id: number;
    start: string;
    team?: Team | null;
    title: string;
}